requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-evtx==0.7.4
xmltodict==0.13.0
Flask-Cors==4.0.0